refactor: :truck: mudança de pasta do formulário de registro
